requirements.txt
pygame
noise
numpy
PyOpenGL
PyOpenGL_accelerate